+++ src/main/java/com/my/relink/domain/user/repository/UserRepository.java
@@ -1,5 +1,6 @@
-package com.my.relink.domain.user;
+package com.my.relink.domain.user.repository;
 
+import com.my.relink.domain.user.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
