+++ src/test/java/com/my/relink/service/UserServiceTest.java
@@ -4,11 +4,11 @@ import com.my.relink.domain.image.EntityType;
 import com.my.relink.domain.image.Image;
 import com.my.relink.domain.image.ImageRepository;
 import com.my.relink.domain.user.User;
-import com.my.relink.domain.user.UserRepository;
-import com.my.relink.dto.user.req.AddressCreateReqDto;
-import com.my.relink.dto.user.req.UserCreateReqDto;
-import com.my.relink.dto.user.resp.UserCreateRespDto;
-import com.my.relink.dto.user.resp.UserInfoRespDto;
+import com.my.relink.domain.user.repository.UserRepository;
+import com.my.relink.controller.user.dto.req.AddressCreateReqDto;
+import com.my.relink.controller.user.dto.req.UserCreateReqDto;
+import com.my.relink.controller.user.dto.resp.UserCreateRespDto;
+import com.my.relink.controller.user.dto.resp.UserInfoRespDto;
 import com.my.relink.ex.BusinessException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
